Added BRCode controller and service. Refactored AuthPayment service method in CommonClass. Moved JUnit tests.
