+++ src/main/java/it/gov/pagopa/payment/dto/brcode/TransactionBRCodeCreationRequest.java
@@ -1,0 +1,17 @@
+package it.gov.pagopa.payment.dto.brcode;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class TransactionBRCodeCreationRequest {
+
+    @NotBlank
+    private String initiativeId;
+}
