+++ src/main/java/it/gov/pagopa/payment/service/payment/BRCodePaymentService.java
@@ -1,0 +1,11 @@
+package it.gov.pagopa.payment.service.payment;
+
+import it.gov.pagopa.payment.dto.AuthPaymentDTO;
+import it.gov.pagopa.payment.dto.brcode.TransactionBRCodeCreationRequest;
+import it.gov.pagopa.payment.dto.brcode.TransactionBRCodeResponse;
+
+public interface BRCodePaymentService {
+
+    TransactionBRCodeResponse createTransaction(TransactionBRCodeCreationRequest trxBRCodeCreationRequest, String userId);
+    AuthPaymentDTO authPayment(String trxCode, String merchantId);
+}
