+++ src/main/java/it/gov/pagopa/payment/service/payment/BRCodePaymentServiceImpl.java
@@ -1,0 +1,23 @@
+package it.gov.pagopa.payment.service.payment;
+
+
+import it.gov.pagopa.payment.dto.AuthPaymentDTO;
+import it.gov.pagopa.payment.dto.brcode.TransactionBRCodeCreationRequest;
+import it.gov.pagopa.payment.dto.brcode.TransactionBRCodeResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+public class BRCodePaymentServiceImpl implements BRCodePaymentService {
+
+    @Override
+    public TransactionBRCodeResponse createTransaction(TransactionBRCodeCreationRequest trxBRCodeCreationRequest, String userId) {
+        return null; //TODO after refactor impl
+    }
+
+    @Override
+    public AuthPaymentDTO authPayment(String trxCode, String merchantId) {
+        return null; //TODO after refactor impl
+    }
+}
